requirements.txt
altgraph==0.17.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pefile==2023.2.7
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.4
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
openai>=1.3.8,<2.0.0
pillow>=9.0.0,<11.0.0
pymongo>=4.0